HR-1695 Introducir más caracteres en Valor Actual. Cherry pick de: BKREC-1652 Aumentar precision a BigDecimal de Bien.valorActual
Former-commit-id: 5efa2fe00bbce96867c469192f78cca53b770597

+++ recovery-nuevoModeloBienes-plugin/src/main/java/es/pfsgroup/plugin/recovery/nuevoModeloBienes/bienes/controller/EditBienController.java
@@ -636,7 +636,7 @@ public class EditBienController {
 			dto.setParticipacion(Integer.parseInt(request
 					.getParameter("oldParticipacion")));
 		if (!Checks.esNulo(request.getParameter("oldValorActual")))
-			dto.setValorActual(new Float(request.getParameter("oldValorActual")));
+			dto.setValorActual(new BigDecimal(request.getParameter("oldValorActual")));
 		if (!Checks.esNulo(request.getParameter("oldImporteCargas")))
 			dto.setImporteCargas(new Float(request
 					.getParameter("oldImporteCargas")));
@@ -659,7 +659,7 @@ public class EditBienController {
 		if (!Checks.esNulo(request.getParameter("tipoBien")))
 			dto.setTipoBien(request.getParameter("tipoBien"));
 		if (!Checks.esNulo(request.getParameter("valorActual")))
-			dto.setValorActual(new Float(request.getParameter("valorActual")));
+			dto.setValorActual(new BigDecimal(request.getParameter("valorActual")));
 		if (!Checks.esNulo(request.getParameter("importeCargas")))
 			dto.setImporteCargas(new Float(request
 					.getParameter("importeCargas")));
